My Arms were short to box with god
Hack up a BSP based solution. Collisions can be derived from plane data we get in BSP form much like we did with maps; but the structure is a bit crazier.

Though we get free occlusion which is nice...

This is proof of concept and on a separate path from QMap; there are still reasonable uses for that. The main work will be finding a way for both to play nice together; right now QMap and BSPMap can't share entities.